+++ package.json
@@ -27,10 +27,6 @@
     "buildCont": "cd command && npm run buildCont"
   },
   "dependencies": {
-    "@tensorflow-models/coco-ssd": "^2.2.2",
-    "@tensorflow/tfjs-converter": "^3.11.0",
-    "@tensorflow/tfjs-core": "^3.11.0",
-    "@tensorflow/tfjs-node": "3.3.0",
     "command": "file:./command",
     "cookie-parser": "^1.4.5",
     "dotenv": "^10.0.0",
